implement reviews: move resolveFileName to viewModel
