+++ feature/course-view/src/debug/kotlin/de/tum/informatics/www1/artemis/native_app/feature/courseview/MessagingScreenshots.kt
@@ -257,7 +257,7 @@ fun `Metis - Conversation Channel`() {
                                 onClickViewPost = {},
                                 onRequestRetrySend = {},
                                 title = "",
-                                onFileSelected = {_, _  ->}
+                                onFileSelected = { _ ->}
                             )
                         }
                     )
