+++ feature/metis/conversation/src/test/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/ReplyTextFieldUiTest.kt
@@ -67,7 +67,7 @@ class ReplyTextFieldUiTest {
                     },
                     updateFailureState = {},
                     title = "TestChat",
-                    onFileSelect = { _, _ -> }
+                    onFileSelected = { _ -> }
                 )
             }
         }
